Rename `xoid/setup` to `xoid` root
