+++ packages/xoid/src/index.tsx
@@ -1,8 +1,10 @@
 import { atom } from './atom'
-import { inject as injectDeprecated, effect as effectDeprecated } from './setup'
-
-export * from './internal/types'
-export { atom }
+import {
+  inject as injectDeprecated,
+  effect as effectDeprecated,
+  createAdapter,
+  setup,
+} from './setup'
 
 /**
  * @deprecated since version 1.0.0beta-12
@@ -24,5 +26,10 @@ export const create = atom
  */
 export default create
 
+export { atom }
 export const inject = injectDeprecated
 export const effect = effectDeprecated
+export { createAdapter, setup }
+
+export type { InjectionKey, EffectCallback } from './setup'
+export * from './internal/types'
